requirements.txt
mnemonic
bip32utils
requests
python-dotenv
pyfiglet
